requirements.txt
appdirs==1.4.4
asgiref==3.5.2
audioread==2.1.9
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
decorator==5.1.1
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
joblib==1.1.0
librosa==0.9.1
llvmlite==0.38.1
numba==0.55.2
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pooch==1.6.0
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.9
pytz==2022.1
requests==2.28.0
resampy==0.2.2
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
SoundFile==0.10.3.post1
sqlparse==0.4.2
threadpoolctl==3.1.0
urllib3==1.26.9
